+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/button
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/ripple@0.0.32
+
 ## 0.0.31
 
 ### Patch Changes