+++ packages/components/image-upload/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jamsr-ui/image-upload
 
+## 0.0.38
+
+### Patch Changes
+
+- fix consoles
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.29
+
+## 0.0.37
+
+### Patch Changes
+
+- image with file
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.28
+
 ## 0.0.36
 
 ### Patch Changes