+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @jamsr-ui/input
 
+## 0.0.35
+
+### Patch Changes
+
+- fix consoles
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.29
+  - @jamsr-ui/button@0.0.32
+
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.28
+  - @jamsr-ui/button@0.0.31
+
 ## 0.0.33
 
 ### Patch Changes